refactor: organize legacy source codes
