+++ src/legacy/generate/cpp.ts
