+++ src/legacy/generate/cpp/assertions.ts
