+++ src/legacy/generate/cpp/headers.test.ts
@@ -2,8 +2,8 @@ import { jest } from "@jest/globals";
 import { createTempDirectory, ITempDirectory } from "create-temp-directory";
 import fs from "node:fs/promises";
 import path from "node:path";
-import { compileCppSource } from "../../compile/cpp.js";
-import { runExecutable } from "../../run.js";
+import { compileCppSource } from "../../../compile/cpp.js";
+import { runExecutable } from "../../../run.js";
 import { generateCppIncludeHeadersCode } from "./headers.js";
 
 jest.retryTimes(10);
