+++ src/legacy/generate/cpp/headers.ts
