+++ src/legacy/generate/cpp/main.ts
