+++ src/legacy/generate/cpp/utility.test.ts
