+++ src/legacy/generate/cpp/utility.ts
