+++ src/legacy/generate/cpp/variable.ts
